+++ latte-client/src/app/components/password/password.component.html
@@ -1,6 +1,6 @@
 @if (control) {
-  <div class="flex items-center gap-1 border-[.1em] border-gray-200 px-2 py-1 rounded-sm focus-within:border-lime-500">
-    <input class="text-base text-neutral-800 grow outline-none" [placeholder]="placeholder? placeholder : 'your-secret'" [id]="id" [type]="type" [formControl]="control">
-    <fa-icon class="text-sm text-neutral-500 cursor-pointer" [title]="type === 'password'? 'show password' : 'hide password'" [icon]="type === 'password'? 'eye' : 'eye-slash'" (click)="show()"></fa-icon>
+  <div class="flex items-center gap-1 border-[.1em] border-neutral-200 rounded-sm  px-2 py-1 focus-within:border-green-600">
+    <input class="w-full text-base text-neutral-900 outline-none" [id]="id" [type]="type" [placeholder]="placeholder" [formControl]="control">
+    <fa-icon class="text-sm text-neutral-400 cursor-pointer" [title]="type === 'password'? 'show password' : 'hide password'" [icon]="type === 'password'? 'eye' : 'eye-slash'" (click)="show()"></fa-icon>
   </div>
 }
