+++ latte-client/src/app/page/home/home.component.css
@@ -3,7 +3,7 @@
 }
 
 .active {
-  background-color: oklch(0.768 0.233 130.85);
+  background-color: oklch(0.723 0.219 149.579);
   color: #fff;
 }
