+++ latte-client/src/app/page/ticket-details/components/activity/activity.component.ts
