+++ latte-client/src/app/page/ticket-details/ticket-details.component.ts
