--- latte-client/src/app/pages/auth/auth.component.html
@@ -1,37 +1,0 @@
-<div class="w-[min(100vw,450px)] h-screen mx-auto flex flex-col items-center justify-center p-4 gap-8">
-  <div class="flex items-end gap-2">
-    <div class="w-[10rem] h-full my-[.1em]">
-      <img class="w-full h-full object-contain" src="asset/icon.svg" alt="logo">
-    </div>
-  </div>
-
-  <div class="w-full">
-    <h1 class="text-3xl text-center font-bold">Sign in to your account</h1>
-
-    <form class="w-full bg-white p-4 mt-4 rounded-lg border-[.1em] border-gray-200" [formGroup]="form" (ngSubmit)="onSubmit()">
-      <div class="flex flex-col gap-1 my-4">
-        <label class="text-sm text-neutral-800" for="email">Email Address</label>
-        <input class="text-base text-neutral-800 border-[.1em] border-gray-200 px-2 py-1 rounded-sm outline-none focus:border-lime-500" placeholder="petergriffin@something.com" id="email" type="email" formControlName="email">
-        @if(formError && (formControls['email'].errors && formControls['email'].errors['required'])) {
-          <span class="text-sm text-red-600">*Required email address</span>
-        }
-        @if(formError && (formControls['email'].errors && formControls['email'].errors['email'])) {
-          <span class="text-sm text-red-600">*Enter valid email address</span>
-        }
-      </div>
-
-      <div class="flex flex-col gap-1 my-4">
-        <label class="text-sm text-neutral-800" for="password">Password</label>
-        <app-password [id]="'password'" formControlName="password"/>
-        @if(formError && (formControls['password'].errors && formControls['password'].errors['required'])) {
-          <span class="text-sm text-red-600">*Required password</span>
-        }
-        @if(formError && (formControls['password'].errors && formControls['password'].errors['minlength'])) {
-          <span class="text-sm text-red-600">*Pasword must have minimum 8 character</span>
-        }
-      </div>
-
-      <button class="bg-lime-500 text-base text-white font-medium rounded-sm border-[.1em] border-lime-500 w-full my-2 px-2 py-1 cursor-pointer hover:bg-lime-600 hover:border-lime-600" type="submit">Sign in</button>
-    </form>
-  </div>
-</div>
