--- latte-client/src/app/pages/home/home.component.html
@@ -1,66 +1,0 @@
-@if(ticketToggle) {
-  <app-ticket-form (cancel)="ticketToggle = false"/>
-}
-
-<div class="w-screen h-screen flex flex-col">
-  <input type="checkbox" id="nav-toggle" class="fixed hidden nav-toggle" #navtoggle>
-
-  <div class="navbar px-4 py-2 border-b-[.1em] border-b-gray-200">
-    <div class="flex items-end">
-      <div class="w-[6.5rem] h-full my-[.1em]">
-        <img class="w-full h-full object-contain" src="asset/icon.svg" alt="logo">
-      </div>
-      
-      <label for="nav-toggle" class="hide">
-        @if (!navtoggle.checked) {
-          <fa-icon class="text-base text-lime-500 cursor-pointer" [icon]="'arrow-right-to-bracket'"></fa-icon>
-        } @else {
-          <fa-icon class="text-base text-lime-500 cursor-pointer transform rotate-90" [icon]="'arrow-right-to-bracket'"></fa-icon>
-        }
-      </label>
-    </div>
-  </div>
-
-  <div class="w-full h-full flex">
-    <div class="bg-white h-full flex flex-col w-[min(100%,280px)] border-r-[.1em] border-r-gray-200 p-4 sidenav">
-      <div class="flex flex-col gap-2">
-        <a class="text-neutral-600 font-medium  p-2 rounded-sm cursor-pointer flex items-center gap-2" (click)="createTicket()">
-          <fa-icon class="text-[.9rem]" [icon]="'pen'"></fa-icon>
-          <span class="text-base">Create Ticket</span>
-        </a>
-  
-        <a class="text-neutral-600 text-base font-medium p-2 rounded-sm cursor-pointer flex items-center gap-2" routerLink="/home/dashboard" routerLinkActive="active">
-          <fa-icon class="text-[.9rem]" [icon]="'house'"></fa-icon>
-          <span>Dashboard</span>
-        </a>
-  
-        <a class="text-neutral-600 text-base font-medium p-2 rounded-sm cursor-pointer flex items-center gap-2" routerLink="/home/tickets" routerLinkActive="active">
-          <fa-icon class="text-[.9rem]" [icon]="'ticket'"></fa-icon>
-          <span>Tickets</span>
-        </a>
-      </div>
-
-      <div class="mt-auto flex flex-col gap-2">
-        @if (isAdmin()) {
-          <span class="text-neutral-800 text-base font-medium px-2 rounded-sm">ADMIN</span>
-          <a class="text-neutral-600 text-base px-2 rounded-sm cursor-pointer" routerLink="/home/users">Users</a>
-        }
-        <a class="text-neutral-600 text-base px-2 rounded-sm cursor-pointer" href="https://github.com/manishdait/latte">Version - <span class="text-sm text-neutral-800 font-bold bg-lime-100 py-1 px-2 rounded-sm">{{ version }}</span></a>
-
-        <div class="flex gap-2 p-2 items-center mt-4 cursor-pointer" routerLink="/home/profile">
-          <div class="w-[2.2rem] h-[2.2rem] rounded-[100%] flex items-center justify-center" [style.background-color]="color(firstname)">
-            <span class="text-base text-white font-bold">{{ firstname.charAt(0) }}</span>
-          </div>
-
-          <div class="flex flex-col gap-1">
-            <span class="text-sm text-neutral-600">{{ firstname }}</span>
-            <span class="text-sm text-neutral-600 font-medium">View Profile</span>
-          </div>
-        </div>
-      </div>
-    </div>
-
-    <div class="grow overflow-y-scroll">
-      <router-outlet></router-outlet>
-    </div>
-</div>
