+++ latte-client/src/app/routes/dashboard/dashboard.component.html
@@ -1,23 +1,23 @@
-<div class="w-full p-4 flex flex-col gap-8">
-  <div class="bg-white p-8 rounded-sm flex flex-col gap-2 border-[.1em] border-gray-200">
-    <h1 class="text-3xl font-extrabold ">{{ greet() + ', ' + firstname }}</h1>
+<div class="w-full flex flex-col gap-8 p-4">
+  <div class="bg-neutral-50 flex flex-col gap-4 border-[.1em] border-neutral-200 rounded-sm p-4">
+    <h1 class="text-3xl text-neutral-900 font-bold">{{ greet() + ', ' + firstname }}</h1>
 
     <div class="flex items-center gap-2">
-      <div class="w-[3rem] h-[3rem] rounded-[100%] flex items-center justify-center" [style.background-color]="color(firstname)">
+      <div class="w-[2.5rem] h-[2.5rem] flex items-center justify-center rounded-[100%]" [style.background-color]="color(firstname)">
         <span class="text-base font-bold text-white">{{ firstname.charAt(0) }}</span>
       </div>
 
-      <span class="text-base">{{ firstname }}</span>
+      <span class="text-base text-neutral-900">{{ firstname }}</span>
     </div>
   </div>
 
-  <div class="flex gap-8 flex-wrap">
-    <div class="bg-white p-8 rounded-sm flex flex-col gap-2 w-[min(100%,400px)] border-[.1em] border-gray-200">
+  <div class="flex flex-wrap gap-8">
+    <div class="w-[min(100%,400px)] flex flex-col gap-4 bg-neutral-50 text-neutral-900 border-[.1em] border-neutral-200 rounded-sm p-4">
       <span class="text-base">Open Tickets</span>
       <span class="text-4xl font-bold">{{ (openCount$ | async) }}</span>
     </div>
 
-    <div class="bg-white p-8 rounded-sm flex flex-col gap-2 w-[min(100%,400px)] border-[.1em] border-gray-200">
+    <div class="w-[min(100%,400px)] flex flex-col gap-4 bg-neutral-50 text-neutral-900 border-[.1em] border-neutral-200 rounded-sm p-4">
       <span class="text-base">Completed Tickets</span>
       <span class="text-4xl font-bold">{{ (closeCount$ | async) }}</span>
     </div>
