+++ latte-client/src/styles.css
@@ -1,14 +1,10 @@
 @import "tailwindcss";
 
-body {
-  background-color: #fff;
-}
-
 html {
   font-size: 1rem;
 }
 
-@media screen and (max-width: 700px) {
+@media screen and (max-width: 600px) {
   html {
     font-size: .9rem;
   }
